Ticket #446: Replaced long long with int

+++ src/db/mysql/sociConversions.h
@@ -109,7 +109,7 @@ namespace soci
             file.DEST_SPACE_TOKEN   = v.get<std::string>("space_token");
             file.REASON             = v.get<std::string>("reason", "");
 
-            long long size = v.get<long long>("filesize", 0);
+            unsigned size = static_cast<unsigned>(v.get<int>("filesize", 0));
             std::ostringstream str;
             str << size;
             file.FILESIZE = str.str();
